# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cartoon Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/index": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar produtos",
                "description": "Retorna uma lista de produtos",
                "operationId": "2aa9a961d000b298a8e18cd347c5ce24",
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/cartoon/indexOfProductForUser": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar produtos",
                "description": "Retorna uma lista de produtos",
                "operationId": "fff57a495dc9a2facbc2673ff1e4c51a",
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/cartoon/indexMoreSold": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista os produtos mais vendidos",
                "description": "Lista os produtos mais vendidos",
                "operationId": "9edf0e6c5ad88f6a4d0c8987edafcc32",
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/product/show": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product information",
                "operationId": "cf6632144ae9cac301839b1a925f29b7",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "ID of the product to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "59"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "John"
                                        },
                                        "category": {
                                            "type": "object"
                                        },
                                        "stock": {
                                            "type": "object"
                                        },
                                        "photos": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/storage": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Armazenar um novo produto",
                "description": "Armazena um novo produto no sistema.",
                "operationId": "a7dae1ae99031ad98ed7c27e0c13f4c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "Tereuba"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock_min": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "stock_max": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "image[0]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image[1]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image[2]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image[3]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product/update": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualizar um produto",
                "description": "Atualizar um novo produto no sistema.",
                "operationId": "e1ef0c392bc281c0b1cca48adb5a23e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Agenda"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "Tereuba"
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "stock_min": {
                                        "type": "string",
                                        "example": 100
                                    },
                                    "stock_max": {
                                        "type": "string",
                                        "example": 100
                                    },
                                    "stock_current": {
                                        "type": "string",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atualizado criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token sem autorização."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/destroy": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deletar um usuário",
                "description": "Atualizar um novo usuário no sistema.",
                "operationId": "67cfda5d0a8413107c20aae9ee29cb2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token sem autorização."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [
                            "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hY2NvdW50XC9sb2dpbiIsImlhdCI6MTY3OTI0NDIxNSwiZXhwIjoxNjc5MjQ3ODE1LCJuYmYiOjE2NzkyNDQyMTUsImp0aSI6IjRsdkhYUEZzeWJFYTAyMkUiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.MumzXQs_Rfa2UMSEUDJtr-q4sOr--DclRLA9FUA2P5A"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model",
                "xml": {
                    "name": "Product"
                }
            },
...
        },
...
    }
}